Split AST domain & identity functions
